Makefile: fix ccache not being used by default, and try to automatically detect it
Someone previously made CXX be set conditionally to 'ccache g++' when
not previously set, but CXX is always set, being a built-in variable and
all.

Now we introduce two options to the Makefile:

 * CCACHE: the ccache binary that should be used when USE_CCACHE is set
   to 'yes'. Defaults to 'ccache'.
 * USE_CCACHE: whether to use the CCACHE binary during builds. Defaults
   to 'yes', but see below.

USE_CCACHE is unset at the start when the CCACHE path does not seem to
point to the ccache binary (via which(1)), thus disabling ccache support
when not present.

People using wrapper scripts may want to stop explicitly passing
'ccache' in the CXX variable, otherwise it will appear twice in the
compiler command line.