Made some questionable optimizations to sound_source.cfg.
You must be in the editor to get the check on sound_file in _sound_defaults.
Cached _sound_volume_scale; it will only be calculated once when the object is instantiated. FFL isn't as proactive about caching static values as it could be. (Am I missing something, like CONSTS_IN_CAPS?)