Fixed level.player.position_predictor which was getting stuck with a stale ref to an old Frogatto after a respawn.
This fixes a few enemies not attacking, such as the spiders. Due to a quirk
in how the position_predictor worked, this meant that the farther away you
were from the point of death, the more inaccurate the predicted position
would be. So the enemy you died to would still work fine, but all the rest
of the enemies - after your first death on the level - would be unable to
accurately predict where you were after.